+++ backend/package.json
@@ -6,7 +6,7 @@
   "type": "commonjs",
   "scripts": {
     "start": "nodemon server.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
@@ -14,17 +14,24 @@
   "devDependencies": {
     "@types/express": "^4.17.17",
     "@types/express-session": "^1.17.7",
+    "@types/jest": "^28.0.0",
     "@types/node": "^20.4.2",
+    "jest": "^28.0.0",
     "nodemon": "^3.0.1",
+    "ts-jest": "^28.0.0",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6"
   },
   "dependencies": {
     "@prisma/client": "^5.1.1",
+    "@types/superagent": "^4.1.18",
+    "@types/supertest": "^2.0.12",
     "bcrypt": "^5.1.1",
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
-    "prisma": "^5.1.1"
+    "prisma": "^5.1.1",
+    "superagent": "^8.1.2",
+    "supertest": "^6.3.3"
   }
 }
